+++ app/src/main/java/com/cmput301w20t10/uberapp/util/HistoryAdapter.java
@@ -80,7 +80,7 @@ public class HistoryAdapter extends BaseAdapter {
         // get the easily accessable ride request information
         RideRequest request = rideHistory.get(position);
         Date time = request.getTimestamp();
-        float offer = request.getFareOffer();
+        int offer = request.getFareOffer();
 
         // get reference to the UI elements
         TextView fareView = view.findViewById(R.id.rideFare);
@@ -91,7 +91,7 @@ public class HistoryAdapter extends BaseAdapter {
         // write the easily accessible information
         DateFormat dateFormat = new SimpleDateFormat("yy-MM-dd hh:mm a");
         dateView.setText(dateFormat.format(time));
-        fareView.setText("$"+String.format("%.2f", offer));
+        fareView.setText("$"+String.format("%d", offer));
         statusText.setText(String.valueOf(request.getState()));
 
         if (user instanceof Rider) {
